Moiving components in file tree to reorganize
