+++ src/applications/personalization/profile/components/contact-information/email-addresses/EmailAddressNotification.jsx
